+++ airbyte-cdk/bulk/toolkits/load-db/src/main/kotlin/io/airbyte/cdk/load/orchestration/db/DestinationInitialStatusGatherer.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.load.orchestration
+package io.airbyte.cdk.load.orchestration.db
 
 import io.airbyte.cdk.load.command.DestinationStream
 
@@ -24,6 +24,6 @@ interface DestinationInitialStatus
  */
 fun interface DestinationInitialStatusGatherer<InitialStatus : DestinationInitialStatus> {
     suspend fun gatherInitialStatus(
-        streams: Map<DestinationStream, Pair<TableNames, ColumnNameMapping>>,
+      streams: Map<DestinationStream, Pair<TableNames, ColumnNameMapping>>,
     ): Map<DestinationStream, InitialStatus>
 }
