+++ airbyte-cdk/bulk/toolkits/load-db/src/main/kotlin/io/airbyte/cdk/load/orchestration/db/direct_load_table/DirectLoadSqlGenerator.kt
@@ -2,6 +2,6 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.load.orchestration.direct_load_table
+package io.airbyte.cdk.load.orchestration.db.direct_load_table
 
 interface DirectLoadSqlGenerator {}
