+++ airbyte-cdk/bulk/toolkits/load-db/src/main/kotlin/io/airbyte/cdk/load/orchestration/db/legacy_typing_deduping/TypingDedupingUtil.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.load.orchestration.legacy_typing_deduping
+package io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping
 
 import kotlin.math.max
