+++ airbyte-cdk/bulk/toolkits/load-db/src/main/kotlin/io/airbyte/cdk/load/orchestration/db/legacy_typing_deduping/TypingDedupingWriter.kt
@@ -2,13 +2,13 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.load.orchestration.legacy_typing_deduping
+package io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping
 
 import io.airbyte.cdk.load.command.DestinationStream
-import io.airbyte.cdk.load.orchestration.ColumnNameMapping
-import io.airbyte.cdk.load.orchestration.DestinationHandler
-import io.airbyte.cdk.load.orchestration.DestinationInitialStatusGatherer
-import io.airbyte.cdk.load.orchestration.TableNames
+import io.airbyte.cdk.load.orchestration.db.ColumnNameMapping
+import io.airbyte.cdk.load.orchestration.db.DestinationHandler
+import io.airbyte.cdk.load.orchestration.db.DestinationInitialStatusGatherer
+import io.airbyte.cdk.load.orchestration.db.TableNames
 import io.airbyte.cdk.load.write.DestinationWriter
 import io.airbyte.cdk.load.write.StreamLoader
 import java.util.concurrent.Executors
@@ -17,13 +17,13 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 
 class TypingDedupingWriter(
-    private val names: Map<DestinationStream, Pair<TableNames, ColumnNameMapping>>,
-    private val stateGatherer:
-        DestinationInitialStatusGatherer<TypingDedupingDestinationInitialStatus>,
-    private val destinationHandler: DestinationHandler,
-    private val rawTableOperations: TypingDedupingRawTableOperations,
-    private val finalTableOperations: TypingDedupingFinalTableOperations,
-    private val disableTypeDedupe: Boolean,
+  private val names: Map<DestinationStream, Pair<TableNames, ColumnNameMapping>>,
+  private val stateGatherer:
+  DestinationInitialStatusGatherer<TypingDedupingDestinationInitialStatus>,
+  private val destinationHandler: DestinationHandler,
+  private val rawTableOperations: TypingDedupingRawTableOperations,
+  private val finalTableOperations: TypingDedupingFinalTableOperations,
+  private val disableTypeDedupe: Boolean,
 ) : DestinationWriter {
     private lateinit var initialStatuses:
         Map<DestinationStream, TypingDedupingDestinationInitialStatus>
